session-stored flag retrieval
